python-server/requirements.txt
fastmcp>=3.0.0
uvicorn[standard]>=0.41.0
pydantic>=2.12.0
httpx[asyncio]>=0.27.0
redis[asyncio]>=5.0.0
